# docs/swagger.yaml
basePath: /
definitions:
  controller.Info:
    properties:
      message:
        example: 消息
        type: string
      ok:
        example: true
        type: boolean
    type: object
  controller.JSONResult:
    properties:
      code:
        example: 0
        type: integer
      data:
        type: object
      info:
        $ref: '#/definitions/controller.Info'
    type: object
info:
  contact: {}
  description: 网页版kubectl cp
  title: KubeCp Swagger
  version: "1.0"
paths:
...
  /api/k8s/status:
    get:
      description: 获取pod中container状态
      parameters:
      - in: query
        name: deployment
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: deployment
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: deployment
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: deployment
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.JSONResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.JSONResult'
      summary: PodStatus
      tags:
      - Kubernetes
  /api/k8s/terminal:
    get:
      description: pod 中容器的终端
      parameters:
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: shell
        type: string
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: shell
        type: string
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: shell
        type: string
      - in: query
        name: container
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pods
        required: true
        type: string
      - in: query
        name: shell
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.JSONResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.JSONResult'
      summary: Container terminal
      tags:
      - Kubernetes
...
swagger: "2.0"